添加基础类
