+++ MVVMFrame/src/main/java/com/leonyr/mvvm/act/AbActivity.java
@@ -1,4 +1,4 @@
-package com.leonyr.lib.mvvm.act;
+package com.leonyr.mvvm.act;
 
 import android.content.Context;
 import android.os.Bundle;
@@ -7,7 +7,7 @@ import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.view.Window;
 
-import com.leonyr.lib.mvvm.vm.LViewModel;
+import com.leonyr.mvvm.vm.LViewModel;
 
 /**
  * ==============================================================
