+++ MVVMFrame/src/main/java/com/leonyr/mvvm/act/AbBindActivity.java
@@ -1,4 +1,4 @@
-package com.leonyr.lib.mvvm.act;
+package com.leonyr.mvvm.act;
 
 import android.content.Context;
 import android.databinding.DataBindingUtil;
@@ -8,7 +8,7 @@ import android.support.annotation.LayoutRes;
 import android.support.v7.app.AppCompatActivity;
 import android.view.Window;
 
-import com.leonyr.lib.mvvm.vm.LViewModel;
+import com.leonyr.mvvm.vm.LViewModel;
 
 /**
  * ==============================================================
