+++ MVVMFrame/src/main/java/com/leonyr/mvvm/frag/AbBindBottomDialogFragment.java
@@ -1,23 +1,17 @@
-package com.leonyr.lib.mvvm.frag;
+package com.leonyr.mvvm.frag;
 
-import android.app.Dialog;
 import android.content.Context;
 import android.databinding.DataBindingUtil;
 import android.databinding.ViewDataBinding;
 import android.os.Bundle;
 import android.support.annotation.LayoutRes;
-import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.support.v4.app.DialogFragment;
-import android.util.DisplayMetrics;
+import android.support.design.widget.BottomSheetDialogFragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.Window;
-import android.view.WindowManager;
 
-import com.leonyr.lib.R;
-import com.leonyr.lib.mvvm.vm.LViewModel;
+import com.leonyr.mvvm.vm.LViewModel;
 
 /**
  * ==============================================================
@@ -27,7 +21,7 @@ import com.leonyr.lib.mvvm.vm.LViewModel;
  * (C) Copyright LeonyR Corporation 2014 All Rights Reserved.
  * ==============================================================
  */
-public abstract class AbBindDialogFragment<VM extends LViewModel, B extends ViewDataBinding> extends DialogFragment {
+public abstract class AbBindBottomDialogFragment<VM extends LViewModel, B extends ViewDataBinding> extends BottomSheetDialogFragment {
 
     protected String TAG;
     private VM VModel;
@@ -53,33 +47,11 @@ public abstract class AbBindDialogFragment<VM extends LViewModel, B extends View
         if (parent != null) {
             parent.removeView(binding.getRoot());
         }
-
-        getDialog().setCanceledOnTouchOutside(false);
-
         return binding.getRoot();
     }
 
     public B Binding() {
         return binding;
-    }
-
-    protected static float WIDTH_RATIO = 0.85f;
-
-    @Override
-    public void onResume() {
-        Window window = getDialog().getWindow();
-        if (window != null) {
-            DisplayMetrics dm = new DisplayMetrics();
-            window.getWindowManager().getDefaultDisplay().getMetrics(dm);
-            window.setLayout((int) (dm.widthPixels * WIDTH_RATIO), WindowManager.LayoutParams.WRAP_CONTENT);
-        }
-        super.onResume();
-    }
-
-    @NonNull
-    @Override
-    public Dialog onCreateDialog(Bundle savedInstanceState) {
-        return new Dialog(getActivity(), R.style.CommonDialogStyle);
     }
 
     public VM getVModel() {
