+++ MVVMFrame/src/main/java/com/leonyr/mvvm/frag/AbDialogFragment.java
@@ -1,15 +1,21 @@
-package com.leonyr.lib.mvvm.frag;
+package com.leonyr.mvvm.frag;
 
+import android.app.Dialog;
 import android.content.Context;
 import android.os.Bundle;
 import android.support.annotation.LayoutRes;
+import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+import android.support.v4.app.DialogFragment;
+import android.util.DisplayMetrics;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.view.Window;
+import android.view.WindowManager;
 
-import com.leonyr.lib.mvvm.vm.LViewModel;
+import com.leonyr.mvvm.R;
+import com.leonyr.mvvm.vm.LViewModel;
 
 /**
  * ==============================================================
@@ -19,14 +25,14 @@ import com.leonyr.lib.mvvm.vm.LViewModel;
  * (C) Copyright LeonyR Corporation 2014 All Rights Reserved.
  * ==============================================================
  */
-public abstract class AbFragment<VM extends LViewModel> extends Fragment {
+public abstract class AbDialogFragment<VM extends LViewModel> extends DialogFragment {
 
     protected String TAG;
     private VM VModel;
     protected Context mCtx;
     protected View rootView;
 
-    public AbFragment(){
+    public AbDialogFragment(){
         getLifecycle().addObserver(new FragmentObserver());
     }
 
@@ -49,11 +55,31 @@ public abstract class AbFragment<VM extends LViewModel> extends Fragment {
         if (parent != null) {
             parent.removeView(rootView);
         }
+        getDialog().setCanceledOnTouchOutside(false);
         return rootView;
     }
 
     public View getRootView() {
         return rootView;
+    }
+
+    protected static float WIDTH_RATIO = 0.85f;
+
+    @Override
+    public void onResume() {
+        Window window = getDialog().getWindow();
+        if (window != null) {
+            DisplayMetrics dm = new DisplayMetrics();
+            window.getWindowManager().getDefaultDisplay().getMetrics(dm);
+            window.setLayout((int) (dm.widthPixels * WIDTH_RATIO), WindowManager.LayoutParams.WRAP_CONTENT);
+        }
+        super.onResume();
+    }
+
+    @NonNull
+    @Override
+    public Dialog onCreateDialog(Bundle savedInstanceState) {
+        return new Dialog(getActivity(), R.style.CommonDialogStyle);
     }
 
     public VM getVModel() {
