+++ MVVMFrame/src/main/java/com/leonyr/mvvm/frag/FragmentObserver.java
@@ -1,4 +1,4 @@
-package com.leonyr.lib.mvvm.frag;
+package com.leonyr.mvvm.frag;
 
 import android.arch.lifecycle.Lifecycle;
 import android.arch.lifecycle.LifecycleObserver;
