+++ MVVMFrame/src/main/java/com/leonyr/mvvm/net/RxSchedulers.java
@@ -1,4 +1,4 @@
-package com.leonyr.lib.mvvm.net;
+package com.leonyr.mvvm.net;
 
 import io.reactivex.Observable;
 import io.reactivex.ObservableSource;
